ft: installed vite-tsconfig-paths, refactor: changed some paths to absolute, fix: changes server address
