+++ package.json
@@ -18,7 +18,8 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
     "react-redux": "^8.0.5",
-    "react-select": "^5.7.0"
+    "react-select": "^5.7.0",
+    "vite-tsconfig-paths": "^4.0.3"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.191",
