+++ src/components/modules/left/Left.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react'
-import { useGetStationsQuery } from '../../store/api/giosApi'
-import { useAppDispatch } from '../../store/hooks'
-import { selectStation } from '../../store/slice/stationSlice'
-import { Maybe, Station } from '../../types'
-import { Handler } from '../common/Handler'
+import { useGetStationsQuery } from 'store/api/giosApi'
+import { useAppDispatch } from 'store/hooks'
+import { selectStation } from 'store/slice/stationSlice'
+import { Maybe, Station } from 'types'
+import { Handler } from 'components/common/Handler'
 import { getClosestStation } from './location'
 import { SelectStationContainer } from './select/SelectStationContainer'
