+++ src/main.tsx
@@ -1,13 +1,13 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { App } from './components/App'
 import { Provider } from 'react-redux'
-import { store } from './store/store'
+import { App } from 'components/App'
+import { store } from 'store/store'
 import './main.css'
 
-const root = document.getElementById('root')
+const root = document.getElementById('root')!
 
-createRoot(root!).render(
+createRoot(root).render(
   <StrictMode>
     <Provider store={store}>
       <App />
